refactor: move Button component to shared/ui and tests to component folder
- Move Button from features/button to shared/ui/button
- Move test file to shared/ui/button/__tests__/button.test.tsx
- Move button-variants.ts to shared/ui/button/
- Update all import paths
