+++ apps/sdui-template-component/package.json
@@ -105,4 +105,3 @@
     "url": "https://github.com/lodado/sdui-template/issues"
   }
 }
-
