+++ apps/sdui-template-component/rollup.config.mjs
@@ -2,4 +2,3 @@ import { defaultConfig } from 'rollup-config/rollup.config.mjs'
 
 const config = defaultConfig()
 export default config
-
