+++ apps/sdui-template-component/src/app/index.ts
@@ -1,7 +1,6 @@
-// Feature exports
-export { Button } from '../features/button'
-export type { ButtonProps, ButtonVariant, ButtonSize } from '../features/button'
+// Shared UI components
+export { Button } from '../shared/ui/button'
+export type { ButtonProps, ButtonVariant, ButtonSize } from '../shared/ui/button'
 
 // Shared utilities (if needed for external use)
 export { cn } from '../shared/lib/cn'
-
