--- apps/sdui-template-component/src/features/button/index.ts
@@ -1,3 +1,0 @@
-export { Button } from './ui/Button'
-export type { ButtonProps, ButtonVariant, ButtonSize } from './model/types'
-
