+++ apps/sdui-template-component/src/shared/ui/button/__tests__/button.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import { Button } from '../../features/button'
+import { Button } from '../Button'
 
 describe('Button', () => {
   describe('as is: Button with onClick handler', () => {
