+++ apps/sdui-template-component/src/shared/ui/button/index.ts
@@ -1,0 +1,3 @@
+export { Button } from './Button'
+export type { ButtonProps, ButtonVariant, ButtonSize } from './types'
+
